requirements.txt
astroid==3.0.1
autopep8==2.0.4
click==8.1.7
dill==0.3.7
docstring-to-markdown==0.13
flake8==6.1.0
importlib-metadata==6.8.0
isort==5.12.0
jedi==0.19.1
mccabe==0.7.0
numpy==1.26.2
packaging==23.2
parso==0.8.3
platformdirs==4.0.0
pluggy==1.3.0
pycodestyle==2.11.1
pydocstyle==6.3.0
pyflakes==3.1.0
pylint==3.0.2
PyQt5==5.15.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyqt5-tools==5.15.9.3.3
pyqtgraph==0.13.3
pyserial==3.5
PySide6==6.6.0
PySide6-Addons==6.6.0
PySide6-Essentials==6.6.0
python-dotenv==1.0.0
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.9.0
pytoolconfig==1.2.6
pyuic5-tool==0.0.1
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
rope==1.11.0
shiboken6==6.6.0
snowballstemmer==2.2.0
tomli==2.0.1
tomlkit==0.12.2
ujson==5.8.0
whatthepatch==1.0.5
yapf==0.40.2
zipp==3.17.0
